+++ src/test/java/nextstep/member/acceptance/MemberAcceptanceTest.java
@@ -1,13 +1,14 @@
-package nextstep.subway.acceptance;
+package nextstep.member.acceptance;
 
-import io.restassured.response.ExtractableResponse;
-import io.restassured.response.Response;
+import static nextstep.member.MemberSteps.*;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 
-import static nextstep.subway.acceptance.MemberSteps.*;
-import static org.assertj.core.api.Assertions.assertThat;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import nextstep.common.AcceptanceTest;
 
 class MemberAcceptanceTest extends AcceptanceTest {
     public static final String EMAIL = "email@email.com";
@@ -36,9 +37,9 @@ class MemberAcceptanceTest extends AcceptanceTest {
         // when
         ExtractableResponse<Response> 회원_생성_응답 = 회원_생성_요청(EMAIL, PASSWORD, AGE);
         final String 접근_토큰 = 로그인_되어_있음(EMAIL, PASSWORD);
-        ExtractableResponse<Response> 회원_정보_응답 = 내_회원_정보_조회_요청(접근_토큰);
+        ExtractableResponse<Response> 회원_정보_응답 = 토큰_기반_내_회원_정보_조회_요청(접근_토큰);
         ExtractableResponse<Response> 회원_수정_응답 = 회원_정보_수정_요청(접근_토큰, "new" + EMAIL, "new" + PASSWORD, AGE);
-        ExtractableResponse<Response> 회원_삭제_응답 = 회원_삭제_요청(접근_토큰);
+        ExtractableResponse<Response> 회원_삭제_응답 = 토근_기반_회원_삭제_요청(접근_토큰);
 
         // then
         응답_확인(회원_생성_응답, HttpStatus.CREATED);
